Add scripts for configuring video / network
They are meant to be used through a dedicated UI.  Run without
arguments, these scripts output the supported devices they can manage
(one per line).  Running the same script and passing it a device enable
it's support.

For now, we provide a script that allows configuring the network (using
DHCP on an Ethernet card) and the video output.